requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
zipp==3.8.1
